refactor: サイドバー関連の変更を除外（別PRで対応）
